reviewer/requirements.txt
aiohttp==3.9.1
PyGithub==2.1.1
python-gitlab==4.4.0
google-genai==1.0.0